Rename cmd_vel_mux_nodelet to cmd_vel_mux.
